HelloWorld 01 Exercise
This is my first Exercise doing javascript practical
first time. when i started running the test it was
giving me test failed i waited and looked at the test
what is it expecting from me it was expecting helloworld
then i came to helloWorld.js and changed the empty strings
to "Hello, World" and then the test passed.

Changes to be committed:
modified:   helloWorld.js